docker/my-python-app/requirements.txt
flask
requests  # Используется только эта библиотека
numpy
mysql-connector-python  # Подключаемся к MySQL

